# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SolveX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /solutions:
    post:
      operationId: submitSolution
      tags: [solutions]
      summary: Submit a solution (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolutionInput"
      responses:
        "201":
          description: Solution submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
  /solutions/problem/{problemId}:
    get:
      operationId: getSolutionsByProblem
      tags: [solutions]
      summary: Get solutions for a problem
      parameters:
        - name: problemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Solutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Solution"
  /solutions/{id}/verify:
    post:
      operationId: verifySolution
      tags: [solutions]
      summary: AI-verify a solution (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"

  # ── Escrow ────────────────────────────────────────────────────────────
  /escrow:
    post:
      operationId: createEscrow
      tags: [escrow]
      summary: Create escrow for a problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscrowInput"
      responses:
        "201":
          description: Escrow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escrow"
  /escrow/payment-intent:
    post:
      operationId: createPaymentIntent
      tags: [escrow]
      summary: Create Stripe payment intent for escrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentIntentInput"
      responses:
        "200":
          description: Payment intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentResult"
  /escrow/problem/{problemId}:
    get:
      operationId: getEscrowByProblem
      tags: [escrow]
      summary: Get escrow for a problem
      parameters:
        - name: problemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Escrow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escrow"

  # ── Crawler ───────────────────────────────────────────────────────────
  /crawler/run:
    post:
      operationId: runCrawler
      tags: [crawler]
      summary: Run AI crawler (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrawlerInput"
      responses:
        "200":
          description: Crawler results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerResult"
  /crawler/crawled:
    get:
      operationId: listCrawledProblems
      tags: [crawler]
      summary: List crawled problems (admin only)
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Crawled problems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrawledProblem"
  /crawler/import:
    post:
      operationId: importCrawledProblem
      tags: [crawler]
      summary: Import a crawled problem (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportInput"
      responses:
        "200":
          description: Imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  # ── Notifications ─────────────────────────────────────────────────────
...
  /owner/audit-log:
    get:
      operationId: getAuditLog
      tags: [owner]
      summary: Get audit log (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: eventType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
...
  /owner/withdraw:
    post:
      operationId: initiateOwnerWithdrawal
      tags: [owner]
      summary: Initiate owner withdrawal (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerWithdrawInput"
      responses:
        "200":
          description: Withdrawal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerWithdrawResult"
...
components:
...
  schemas:
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    Solution:
      type: object
      required: [id, problemId, solverId, content, status]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        solverId:
          type: integer
        content:
          type: string
        status:
          type: string
          enum: [pending, verifying, approved, rejected]
        verificationScore:
          type: ["string", "null"]
        verificationNotes:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    SolutionInput:
      type: object
      required: [problemId, content]
      properties:
        problemId:
          type: integer
        content:
          type: string
          minLength: 20
    VerificationResult:
      type: object
      required: [approved, score]
      properties:
        approved:
          type: boolean
        score:
          type: number
        notes:
          type: string
    Escrow:
      type: object
      required: [id, problemId, clientId, amount, status]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        clientId:
          type: integer
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [pending, held, released, refunded, failed]
        stripePaymentIntentId:
          type: ["string", "null"]
        releasedAt:
          type: ["string", "null"]
        refundedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    EscrowInput:
      type: object
      required: [problemId, amount]
      properties:
        problemId:
          type: integer
        amount:
          type: number
    PaymentIntentInput:
      type: object
      required: [problemId, amount]
      properties:
        problemId:
          type: integer
        amount:
          type: number
    PaymentIntentResult:
      type: object
      required: [clientSecret, paymentIntentId]
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
    CrawlerInput:
      type: object
      properties:
        platforms:
          type: array
          items:
            type: string
        category:
          type: string
        limit:
          type: integer
          default: 10
    CrawlerResult:
      type: object
      required: [found]
      properties:
        found:
          type: integer
        problems:
          type: array
          items:
            $ref: "#/components/schemas/CrawledProblem"
    CrawledProblem:
      type: object
      required: [id, title, description, platform]
      properties:
        id:
          type: integer
        externalId:
          type: string
        platform:
          type: string
        title:
          type: string
        description:
          type: string
        sourceUrl:
          type: string
        suggestedCategory:
          type: string
        suggestedPayment:
          type: ["string", "null"]
        isImported:
          type: boolean
        aiSummary:
          type: ["string", "null"]
        upvotes:
          type: ["integer", "null"]
        crawledAt:
          type: string
    ImportInput:
      type: object
      required: [crawledId, paymentOffer]
      properties:
        crawledId:
          type: integer
        paymentOffer:
          type: number
...
    AuditLogEntry:
      type: object
      required: [id, eventType, status, createdAt]
      properties:
        id:
          type: string
        eventType:
          type: string
        userId:
          type: ["integer", "null"]
        orderId:
          type: ["string", "null"]
        details:
          type: string
        status:
          type: string
        createdAt:
          type: string
...
    OwnerWithdrawInput:
      type: object
      required: [entryIds, withdrawalAddress]
      properties:
        entryIds:
          type: array
          items:
            type: string
        withdrawalAddress:
          type: string
    OwnerWithdrawResult:
      type: object
      required: [success, totalAmount, entryCount]
      properties:
        success:
          type: boolean
        totalAmount:
          type: string
        entryCount:
          type: integer
...